add random color
