+++ src/renderer/src/lib/utils/index.ts
@@ -5,15 +5,6 @@ import { Transition } from '../drawable';
 export * from './generateId';
 export * from './roundPoint';
 
-export const randomColor = (): string => {
-  let result = '';
-  for (let i = 0; i < 6; ++i) {
-    const value = Math.floor(16 * Math.random());
-    result += value.toString(16);
-  }
-  return '#' + result;
-};
-
 export const isPointInRectangle = (rectangle: Rectangle, point: Point) => {
   return (
     rectangle.x <= point.x &&
